Add hashicorp-vault chart in common/
This allows us to simplify any values-* file that has to include vault
and not rely on external chart repos.

Note: We add a hashicorp-vault/update-helm-dependency.sh script which
needs to be run after 'helm dependency update hashicorp-vault' is
called. The reason for this is explained in detail in the README.md
file, but the TLDR is that helm has a bug with subcharts and it won't
remove null values in there and so "host: null" breaks the vault route.

Once either the helm bug is fixed or we switch to OCP 4.11 as a minimum
version and to a vault-helm chart that has subdomains support (unmerged
at this point in time), we can drop this one liner patch.

We explored the approach of patching in vault.server.route.host as a
top-level variable. This was investigated here:
https://github.com/hybrid-cloud-patterns/multicloud-gitops/pull/104

It was discarded because it really special-cases vault, which is
undesirable. It also will need changes everywhere once the bugs are
fixed and also it needs changes in the operator as well.